Create services, controllers, DTO and mappers

+++ Backend/src/main/java/com/ntk/TaskFlow/Repository/TaskRepository.java
@@ -4,5 +4,5 @@ import com.ntk.TaskFlow.Entity.Task;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-public interface TaskRepository extends JpaRepository<Task, Long> {
+public interface TaskRepository extends JpaRepository<Task, Integer> {
 }
